Rename vhdstrenc
