+++ vhdutil.h
